refactor(app): introduce view loading component
