+++ app/portainer/components/ViewLoading/index.ts
@@ -1,0 +1,1 @@
+export { ViewLoading, ViewLoadingAngular } from './ViewLoading';
